New package for non ships
